# Make type Hash = Word64