We need to be more careful with remainder operations, we do them on unsigned Word64s.

I limit the size of bloomfilter to 2^48; which is hopefully enough.

Now the elemCheapHashes is *slower*, this is expected: modulus on 64 bit words is more expensive.

Before

```
Benchmarking baseline ...
(This is the cost of just computing the keys.)
Finished.
Time total:        0.12 seconds
Alloc total:       1.20e9 bytes
Time net:          0.12 seconds
Alloc net:         1.20e9 bytes
Time net per key:  4.89e-9 seconds
Alloc net per key: 48.00 bytes

Benchmarking makeCheapHashes ...
(This includes the cost of hashing the keys, less the cost of computing the keys)
Finished.
Time total:        0.34 seconds
Alloc total:       1.20e9 bytes
Time net:          0.22 seconds
Alloc net:         -752.00 bytes
Time net per key:  8.62e-9 seconds
Alloc net per key: -3.01e-5 bytes

Benchmarking elemCheapHashes ...
(this is the simple one-by-one lookup, less the cost of computing and hashing the keys)
Finished.
Time total:        31.75 seconds
Alloc total:       1.20e9 bytes
Time net:          31.41 seconds
Alloc net:         0.00 bytes
Time net per key:  1.26e-6 seconds
Alloc net per key: 0.00 bytes
```

After

```
Benchmarking baseline ...
(This is the cost of just computing the keys.)
Finished.
Time total:        0.12 seconds
Alloc total:       1.20e9 bytes
Time net:          0.12 seconds
Alloc net:         1.20e9 bytes
Time net per key:  4.68e-9 seconds
Alloc net per key: 48.00 bytes

Benchmarking makeCheapHashes ...
(This includes the cost of hashing the keys, less the cost of computing the keys)
Finished.
Time total:        0.30 seconds
Alloc total:       1.20e9 bytes
Time net:          0.18 seconds
Alloc net:         -752.00 bytes
Time net per key:  7.36e-9 seconds
Alloc net per key: -3.01e-5 bytes

Benchmarking elemCheapHashes ...
(this is the simple one-by-one lookup, less the cost of computing and hashing the keys)
Finished.
Time total:        36.34 seconds
Alloc total:       1.20e9 bytes
Time net:          36.04 seconds
Alloc net:         0.00 bytes
Time net per key:  1.44e-6 seconds
Alloc net per key: 0.00 bytes
```