@phadej 2024-03-14
> It might still be nice to mention it somewhere in a prominent Haddock comment.

Sure, will do in the cleanup. As for now, very many comments and haddocks are simply wrong. I don't want to churn here and there.